moved test directory external to the main directory
