+++ tests/madm/test_topsis.py
@@ -42,40 +42,60 @@ from __future__ import unicode_literals
 # DOC
 # =============================================================================
 
-__doc__ = """Test normalization functionalities"""
+__doc__ = """test topsis methods"""
 
 
 # =============================================================================
 # IMPORTS
 # =============================================================================
 
-import numpy as np
-
 from .. import core
-
-from ...weights import equal
-from ... import Data
+from skcriteria import util
+from skcriteria.madm import topsis
 
 
 # =============================================================================
-# BASE
+# BASE CLASS
 # =============================================================================
 
-class EqualTest(core.SKCriteriaTestCase):
+class TopsisTest(core.SKCriteriaTestCase):
+    mnorm = "vector"
+    wnorm = "sum"
 
     def setUp(self):
-        self.nmtx = np.array([
-            [1, 1],
-            [2, 2]
-        ])
-        self.expected = [0.5, 0.5]
+        # Data From:
+        # Tzeng, G. H., & Huang, J. J. (2011).
+        # Multiple attribute decision making: methods and applications.
+        # CRC press.
+        self.mtx = [
+            [5, 8, 4],
+            [7, 6, 8],
+            [8, 8, 6],
+            [7, 4, 6],
+        ]
+        self.criteria = [util.MAX, util.MAX, util.MAX]
 
-    def test_equal(self):
-        result = equal.equal(self.nmtx)
-        self.assertAllClose(result, self.expected)
+    def test_topsis(self):
+        weights = [.3, .4, .3]
 
-    def test_equal_oop(self):
-        data = Data(self.nmtx, [1, 1])
-        wd = equal.EqualWeights()
-        rdata = wd.decide(data)
-        self.assertAllClose(rdata.weights, self.expected)
+        result = [3, 2, 1, 4]
+        points = [0.5037, 0.6581, 0.7482, 0.3340]
+
+        normdata = self.normalize(self.mtx, self.criteria, weights)
+        rank_result, points_result = topsis.topsis(*normdata)
+
+        self.assertAllClose(points_result, points, atol=1.e-4)
+        self.assertAllClose(rank_result, result)
+
+    def test_topsis_dm(self):
+        dm = topsis.TOPSIS()
+
+        weights = [.3, .4, .3]
+
+        result = [3, 2, 1, 4]
+        points = [0.5037, 0.6581, 0.7482, 0.3340]
+
+        decision = dm.decide(self.mtx, self.criteria, weights)
+
+        self.assertAllClose(decision.e_.closeness, points, atol=1.e-4)
+        self.assertAllClose(decision.rank_, result)
