+++ tests/weights/test_divergence.py
@@ -42,7 +42,7 @@ from __future__ import unicode_literals
 # DOC
 # =============================================================================
 
-__doc__ = """test moora methods"""
+__doc__ = """Test normalization functionalities"""
 
 
 # =============================================================================
@@ -50,67 +50,41 @@ __doc__ = """test moora methods"""
 # =============================================================================
 
 from .. import core
-from ... import util
-from ...madm import wsum
+
+from skcriteria.weights import divergence
+from skcriteria import norm, Data, divcorr
 
 
 # =============================================================================
-# BASE CLASS
+# BASE
 # =============================================================================
 
-class WSumTest(core.SKCriteriaTestCase):
-    mnorm = "sum"
-    wnorm = "sum"
+class DivergenceTest(core.SKCriteriaTestCase):
 
     def setUp(self):
-        # Data From:
-        # KRACKA, M; BRAUERS, W. K. M.; ZAVADSKAS, E. K. Ranking
-        # Heating Losses in a Building by Applying the MULTIMOORA . -
-        # ISSN 1392 – 2785 Inzinerine Ekonomika-Engineering Economics, 2010,
-        # 21(4), 352-359.
-
+        # Data from:
+        # Diakoulaki, D., Mavrotas, G., & Papayannakis, L. (1995).
+        # Determining objective weights in multiple criteria problems:
+        # The critic method. Computers & Operations Research, 22(7), 763-770.
         self.mtx = [
-            [33.95, 23.78, 11.45, 39.97, 29.44, 167.10, 3.852],
-            [38.9, 4.17, 6.32, 0.01, 4.29, 132.52, 25.184],
-            [37.59, 9.36, 8.23, 4.35, 10.22, 136.71, 10.845],
-            [30.44, 37.59, 13.91, 74.08, 45.10, 198.34, 2.186],
-            [36.21, 14.79, 9.17, 17.77, 17.06, 148.3, 6.610],
-            [37.8, 8.55, 7.97, 2.35, 9.25, 134.83, 11.935]
+            [61, 1.08, 4.33],
+            [20.7, 0.26, 4.34],
+            [16.3, 1.98, 2.53],
+            [9, 3.29, 1.65],
+            [5.4, 2.77, 2.33],
+            [4, 4.12, 1.21],
+            [-6.1, 3.52, 2.10],
+            [-34.6, 3.31, 0.98]
         ]
-        self.criteria = [
-            util.MIN, util.MIN, util.MIN, util.MIN,
-            util.MAX, util.MIN, util.MAX]
+        self.nmtx = norm.ideal_point(self.mtx, criteria=[1, 1, 1], axis=0)
+        self.expected = [0.27908306,  0.34092628,  0.37999065]
 
-    def test_mdwsum_with_weights(self):
-        weights = [20, 20, 20, 20, 20, 20, 20]
+    def test_divergence(self):
+        result = divergence.divergence(self.nmtx, divcorr.std)
+        self.assertAllClose(result, self.expected)
 
-        result = [5,  1,  3,  6,  4,  2]
-        points = [-0.1075, -0.0037, -0.0468, -0.1560, -0.0732, -0.0413]
-
-        normdata = self.normalize(self.mtx, self.criteria, weights)
-        rank_result, points_result = wsum.mdwsum(*normdata)
-
-        self.assertAllClose(points_result, points, atol=1.e-3)
-        self.assertAllClose(rank_result, result)
-
-    def test_mdwsum(self):
-        result = [5,  1,  3,  6,  4,  2]
-        points = [-0.7526, -0.026, -0.3273, -1.092, -0.5127, -0.2894]
-
-        normdata = self.normalize(self.mtx, self.criteria, weights=None)
-        rank_result, points_result = wsum.mdwsum(*normdata)
-
-        self.assertAllClose(points_result, points, atol=1.e-3)
-        self.assertAllClose(rank_result, result)
-
-    def test_mdwsum_dm(self):
-        weights = [20, 20, 20, 20, 20, 20, 20]
-
-        result = [5,  1,  3,  6,  4,  2]
-        points = [-0.1075, -0.0037, -0.0468, -0.1560, -0.0732, -0.0413]
-
-        dm = wsum.MDWeightedSum()
-        decision = dm.decide(self.mtx, self.criteria, weights)
-
-        self.assertAllClose(decision.e_.points, points, atol=1.e-3)
-        self.assertAllClose(decision.rank_, result)
+    def test_divergence_oop(self):
+        data = Data(self.mtx, [1, 1, 1])
+        wd = divergence.DivergenceWeights()
+        rdata = wd.decide(data)
+        self.assertAllClose(rdata.weights, self.expected)
